[Merge] LikesController

+++ src/main/java/kr/pe/fourj/controller/LikesController.java
@@ -13,7 +13,9 @@ import org.springframework.web.bind.annotation.RestController;
 import kr.pe.fourj.domain.Course;
 import kr.pe.fourj.domain.Likes;
 import kr.pe.fourj.domain.Student;
+import kr.pe.fourj.dto.CourseDTO;
 import kr.pe.fourj.dto.LikesDTO;
+import kr.pe.fourj.dto.StudentDTO;
 import kr.pe.fourj.repository.CourseRepository;
 import kr.pe.fourj.repository.LikesRepository;
 import kr.pe.fourj.repository.StudentRepository;
@@ -38,7 +40,7 @@ public class LikesController {
 		Student student = studentRepository.findById(dto.getStudentIdx()).get();
 		Course course = courseRepository.findById(dto.getCourseIdx()).get();
 		
-		if( isNotAlreadyCart(student, course)) {
+		if( isNotAlreadyLikes(student, course)) {
 			likesRepository.save(new Likes(student,course));
 			return "저장 성공";
 		}else {
@@ -46,7 +48,7 @@ public class LikesController {
 		}	
 	}
 	
-	private boolean isNotAlreadyCart(Student studentDTO , Course courseDTO) {
+	private boolean isNotAlreadyLikes(Student studentDTO , Course courseDTO) {
 		boolean flag = false;
 		Student student = studentRepository.findById(studentDTO.getIdx()).get();
 		Course course = courseRepository.findById(courseDTO.getIdx()).get();
@@ -59,11 +61,14 @@ public class LikesController {
 	
 	//좋아요 삭제
 	@DeleteMapping("/likes")
-	public String deleteLikes(@RequestBody LikesDTO.Delete dto) {
+	public String deleteLikes(@RequestBody LikesDTO.Get dto, StudentDTO.Get studentDto, CourseDTO.Get courseDto) {
 		System.out.println("좋아요 삭제 시도");
 
-		likesRepository.deleteById(dto.getIdx());
-		return "삭제 성공!";
+		if(dto.getStudentIdx()==studentDto.getIdx() || dto.getCourseIdx()==courseDto.getIdx()) {
+			likesRepository.deleteById(dto.getIdx());
+			return "삭제 성공!";
+		}
+		return "삭제 실패!";
 	}
 	
 	//좋아요 단일 조회
@@ -73,6 +78,7 @@ public class LikesController {
 		System.out.println("--좋아요 단일 검색 시도--");
 		
 		return likesRepository.findById(dto.getIdx()).get();
+		
 //		return likesRepository.findById(id).get();
 	}
